+++ src/chat/components/DirectChatWrapper/index.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import * as S from './styles';
-import * as fromActions from '@/modules/shared/state/actions';
+import * as fromActions from '../../state/actions';
 import { debounced } from '@/helpers/index';
 import axios from 'axios';
