+++ src/chat/state/actions/index.js
@@ -1,0 +1,2 @@
+export * from './actionTypes';
+export * from './actionCreators';
