+++ src/modules/shared/state/actions/actionTypes.js
@@ -1,7 +1,5 @@
 export const actionTypes = {
   TOGGLE_LEFT_SIDEBAR: '[UI] TOGGLE_LEFT_SIDEBAR',
   CLOSE_CHAT_WITH_REDIRECT: '[UI] CLOSE_CHAT_WITH_REDIRECT',
-  SET_DIRECT_CHAT_MODE: '[UI] SET_DIRECT_CHAT_MODE',
-  SET_GLOBAL_CHAT_MODE: '[UI] SET_GLOBAL_CHAT_MODE',
 };
