+++ src/modules/shared/state/reducers/index.js
@@ -15,16 +15,6 @@ export const uiReducer = (state = initialState, action) => {
         ...state,
         leftSidebarOpened: action.visible,
       };
-    case actionTypes.SET_GLOBAL_CHAT_MODE:
-      return {
-        ...state,
-        chatMode: 'GLOBAL',
-      };
-    case actionTypes.SET_DIRECT_CHAT_MODE:
-      return {
-        ...state,
-        chatMode: action.chatMode,
-      };
     default:
       return state;
   }
